requirements.txt
Flask==2.3.2
Flask-Cors==4.0.1
Flask-Login==0.6.2
Flask-Migrate== 4.0.4
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.6.0
gunicorn==21.2.0
werkzeug==2.3.7
python-dotenv==0.19.1
eventlet
google-generativeai